Moved the StudentMadagascar.kt to com.example.kotlinplayground foulder, rename class Student to Student in StudentMadagascar.kt , added finctions to StudentMadagascar.kt
